Moved firebase.tsx to services folder
