+++ src/app/ForgotPassword.tsx
@@ -9,7 +9,7 @@ import {
 } from "react-native";
 import Button from "../components/Button";
 import { sendPasswordResetEmail } from "firebase/auth";
-import { auth } from "../firebase";
+import { auth } from "../services/firebase";
 
 export default function ForgotPasswordScreen() {
   const [email, setEmail] = useState("");
